+++ PRUF_TRUFFLE/contracts/PRUF_CORE_MAL.sol
@@ -49,7 +49,7 @@ contract CORE_MAL is PullPayment, BASIC {
     /*
      * @dev retrieves costs from Storage and returns Costs struct
      */
-    function getCost(uint256 _assetClass) internal returns (Costs memory) {
+    function getCost(uint32 _assetClass) internal returns (Costs memory) {
         //^^^^^^^checks^^^^^^^^^
 
         Costs memory cost;
@@ -76,8 +76,8 @@ contract CORE_MAL is PullPayment, BASIC {
     function createRecord(
         bytes32 _idxHash,
         bytes32 _rgtHash,
-        uint256 _assetClass,
-        uint256 _countDownStart
+        uint32 _assetClass,
+        uint32 _countDownStart
     ) internal {
         uint256 tokenId = uint256(_idxHash);
         AC memory AC_info = getACinfo(_assetClass);
@@ -151,7 +151,7 @@ contract CORE_MAL is PullPayment, BASIC {
 
     //--------------------------------------------------------------------------------------Payment internal functions
 
-    function deductNewRecordCosts(uint256 _assetClass) internal whenNotPaused {
+    function deductNewRecordCosts(uint32 _assetClass) internal whenNotPaused {
         //^^^^^^^checks^^^^^^^^^
         Invoice memory pricing;
         //^^^^^^^effects^^^^^^^^^
@@ -165,7 +165,7 @@ contract CORE_MAL is PullPayment, BASIC {
         //^^^^^^^interactions^^^^^^^^^
     }
 
-    function deductRecycleCosts(uint256 _assetClass, address _oldOwner)
+    function deductRecycleCosts(uint32 _assetClass, address _oldOwner)
         internal
         whenNotPaused
     {
@@ -183,7 +183,7 @@ contract CORE_MAL is PullPayment, BASIC {
         //^^^^^^^interactions^^^^^^^^^
     }
 
-    function deductTransferAssetCosts(uint256 _assetClass)
+    function deductTransferAssetCosts(uint32 _assetClass)
         internal
         whenNotPaused
     {
@@ -200,7 +200,7 @@ contract CORE_MAL is PullPayment, BASIC {
         //^^^^^^^interactions^^^^^^^^^
     }
 
-    function deductCreateNoteCosts(uint256 _assetClass) internal whenNotPaused {
+    function deductCreateNoteCosts(uint32 _assetClass) internal whenNotPaused {
         //^^^^^^^checks^^^^^^^^^
         Invoice memory pricing;
         //^^^^^^^effects^^^^^^^^^
@@ -214,7 +214,7 @@ contract CORE_MAL is PullPayment, BASIC {
         //^^^^^^^interactions^^^^^^^^^
     }
 
-    function deductReMintRecordCosts(uint256 _assetClass)
+    function deductReMintRecordCosts(uint32 _assetClass)
         internal
         whenNotPaused
     {
@@ -231,7 +231,7 @@ contract CORE_MAL is PullPayment, BASIC {
         //^^^^^^^interactions^^^^^^^^^
     }
 
-    function deductChangeStatusCosts(uint256 _assetClass)
+    function deductChangeStatusCosts(uint32 _assetClass)
         internal
         whenNotPaused
     {
@@ -248,7 +248,7 @@ contract CORE_MAL is PullPayment, BASIC {
         //^^^^^^^interactions^^^^^^^^^
     }
 
-    function deductForceModifyCosts(uint256 _assetClass) internal whenNotPaused {
+    function deductForceModifyCosts(uint32 _assetClass) internal whenNotPaused {
         //^^^^^^^checks^^^^^^^^^
         Invoice memory pricing;
         //^^^^^^^effects^^^^^^^^^
